changed test and modified main
